ingestion/coinbase/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
websockets==11.0.3
google-cloud-bigquery==3.11.0
python-dateutil==2.8.2
